// src/types/userUsage.ts
export interface UserUsage {
    customer_count: number;
    total_appointments: number;
    total_services: number;
    has_appointments: boolean;
    has_services: boolean;
    has_work_hours: boolean;
    last_login: string;
    from_affiliate: boolean;
    future_appointments: number;
    total_sms_sent: number;
    date_joined: string;
    total_revenue: number;
}